+++ common/common-core/src/main/java/com/xh/common/core/dto/SysOrgRoleDTO.java
@@ -16,11 +16,11 @@ public class SysOrgRoleDTO extends BaseDTO<Integer> {
     @Schema(title = "数据类型", allowableValues = {"1", "2"}, description = "1：用户，2：用户组")
     private Integer type;
     @Schema(title = "用户id或者用户组的id")
-    private Integer userId;
+    private Long userId;
     @Schema(title = "机构id")
-    private Integer sysOrgId;
+    private Long sysOrgId;
     @Schema(title = "角色id")
-    private Integer sysRoleId;
+    private Long sysRoleId;
     @Schema(title = "机构代码")
     private String orgCode;
     @Schema(title = "机构名称")
